+++ resources/dpemotes/Client/Keybinds.lua
@@ -65,6 +65,12 @@ AddEventHandler("dp:ClientKeybindGetOne", function(key, e)
 	if key == "num4" then emob1 = e keyb1 = "num4" elseif key == "num5" then emob2 = e keyb2 = "num5" elseif key == "num6" then emob3 = e keyb3 = "num6" elseif key == "num7" then emob4 = e keyb4 = "num7" elseif key == "num8" then emob5 = e keyb5 = "num8" elseif key == "num9" then emob6 = e keyb6 = "num9" end
 end)
 
+RegisterNetEvent("dp:ClientKeybindDelOne")
+AddEventHandler("dp:ClientKeybindDelOne", function(key, e)
+    SimpleNotify(Config.Languages[lang]['unbound'].. "~g~"..firstToUpper(key).."~w~ keybind.")
+    if key == "num4" then emob1 = e keyb1 = "num4" elseif key == "num5" then emob2 = e keyb2 = "num5" elseif key == "num6" then emob3 = e keyb3 = "num6" elseif key == "num7" then emob4 = e keyb4 = "num7" elseif key == "num8" then emob5 = e keyb5 = "num8" elseif key == "num9" then emob6 = e keyb6 = "num9" end
+end)
+
 -----------------------------------------------------------------------------------------------------
 ------ Functions and stuff --------------------------------------------------------------------------
 -----------------------------------------------------------------------------------------------------
@@ -101,4 +107,16 @@ function EmoteBindStart(source, args, raw)
     end
 end
 
+function EmoteUnbind(source, args, raw)
+    if #args > 0 then
+        local key = string.lower(args[1])
+        local src = source
+
+        TriggerServerEvent("dp:ServerKeybindDelete", key)
+    else
+        SimpleNotify(Config.Languages[lang]['nokeyspecified'])
+    end
 end
+
+end
+